intake: open embed dialog from template actions; reuse copy helper
