+++ src/features/intake/pages/IntakeDetailPage.tsx
@@ -45,6 +45,7 @@ import { resolveIntakeTitle } from '@/features/intake/utils/intakeTitle';
 import { applyConsultationPatchToMetadata } from '@/shared/utils/consultationState';
 import { DEFAULT_INTAKE_TEMPLATE } from '@/shared/constants/intakeTemplates';
 import type { IntakeTemplate, IntakeFieldDefinition } from '@/shared/types/intake';
+import EmbedCodeBlock from '@/features/intake/components/EmbedCodeBlock';
 
 // ── Template helpers ──────────────────────────────────────────────────────────
 
@@ -919,6 +920,14 @@ export const IntakeDetailPage: FunctionComponent<IntakeDetailPageProps> = ({
                 View form setup
               </Button>
             ) : null}
+            {activeTemplate && (practiceDetails as { slug?: string })?.slug && (
+              <div className="mt-4">
+                <EmbedCodeBlock
+                  practiceSlug={(practiceDetails as { slug?: string }).slug}
+                  templateSlug={activeTemplate.slug}
+                />
+              </div>
+            )}
           </div>
           {enrichmentFields.length > 0 ? (
             <dl className="grid grid-cols-1 gap-5 md:grid-cols-3">
