+++ src/features/intake/pages/IntakeTemplatesPage.tsx
@@ -36,6 +36,7 @@ import { fromMinorUnits, toMinorUnitsValue } from '@/shared/utils/money';
 import { getOnboardingStatusPayload } from '@/shared/lib/apiClient';
 import { STANDARD_FIELD_DEFINITIONS, DEFAULT_INTAKE_TEMPLATE } from '@/shared/constants/intakeTemplates';
 import type { FieldPhase, IntakeFieldDefinition, IntakeTemplate } from '@/shared/types/intake';
+import EmbedCodeBlock, { getEmbedSnippet, getPublicFormUrl, copyTextToClipboard } from '@/features/intake/components/EmbedCodeBlock';
 
 type IntakeTemplatesPageProps = {
   onBack?: () => void;
@@ -72,24 +73,6 @@ function slugify(name: string): string {
     .replace(/^-+|-+$/g, '');
 }
 
-function getEmbedSnippet(practiceSlug: string, templateSlug: string): string {
-  return [
-    '<script',
-    `  src="https://app.blawby.com/widget.js?template=${templateSlug}"`,
-    `  data-practice="${practiceSlug}"`,
-    '  async',
-    '></script>',
-  ].join('\n');
-}
-
-function getPublicFormUrl(practiceSlug: string, templateSlug: string): string {
-  const origin = typeof window !== 'undefined' && window.location?.origin
-    ? window.location.origin
-    : 'https://app.blawby.com';
-  const url = new URL(`/public/${encodeURIComponent(practiceSlug)}`, origin);
-  url.searchParams.set('template', templateSlug);
-  return url.toString();
-}
 
 function parseTemplatesFromMetadata(metadata: Record<string, unknown> | null | undefined): IntakeTemplate[] {
   if (!metadata) return [];
@@ -280,48 +263,7 @@ function moveItem<T>(items: T[], fromIndex: number, toIndex: number): T[] {
   return next;
 }
 
-function copyTextToClipboard(
-  text: string,
-  onSuccess: () => void,
-  onError: (message: string) => void,
-) {
-  if (
-    typeof navigator !== 'undefined' &&
-    navigator.clipboard &&
-    typeof navigator.clipboard.writeText === 'function'
-  ) {
-    navigator.clipboard.writeText(text)
-      .then(onSuccess)
-      .catch((error) => {
-        onError(error instanceof Error ? error.message : 'Could not copy to clipboard.');
-      });
-    return;
-  }
-
-  if (typeof document === 'undefined') {
-    onError('Clipboard is not available in this environment.');
-    return;
-  }
-
-  try {
-    const textarea = document.createElement('textarea');
-    textarea.value = text;
-    textarea.setAttribute('readonly', '');
-    textarea.style.position = 'absolute';
-    textarea.style.left = '-9999px';
-    document.body.appendChild(textarea);
-    textarea.select();
-    const copied = document.execCommand('copy');
-    document.body.removeChild(textarea);
-    if (copied) {
-      onSuccess();
-    } else {
-      onError('Clipboard is not available.');
-    }
-  } catch (error) {
-    onError(error instanceof Error ? error.message : 'Clipboard is not available.');
-  }
-}
+ 
 
 function useDragReorder<T>(items: T[], onReorder: (next: T[]) => void) {
   const dragIndexRef = useRef<number | null>(null);
@@ -406,90 +348,7 @@ function maskStripeAccountId(value?: string | null) {
   return `${value.slice(0, 8)}...${value.slice(-4)}`;
 }
 
-type EmbedCodeBlockProps = {
-  practiceSlug: string;
-  templateSlug: string;
-};
-
-export function EmbedCodeBlock({ practiceSlug, templateSlug }: EmbedCodeBlockProps) {
-  const { showSuccess, showError } = useToastContext();
-  const [copied, setCopied] = useState(false);
-  const [copiedLink, setCopiedLink] = useState(false);
-  const snippet = getEmbedSnippet(practiceSlug, templateSlug);
-  const publicUrl = getPublicFormUrl(practiceSlug, templateSlug);
-
-  const handleCopy = () => {
-    copyTextToClipboard(
-      snippet,
-      () => {
-        setCopied(true);
-        showSuccess('Embed copied', 'The widget snippet is ready to paste.');
-        setTimeout(() => setCopied(false), 2000);
-      },
-      (message) => showError('Copy failed', message),
-    );
-  };
-
-  const handleCopyLink = () => {
-    copyTextToClipboard(
-      publicUrl,
-      () => {
-        setCopiedLink(true);
-        showSuccess('Link copied', 'The public intake link is ready to share.');
-        setTimeout(() => setCopiedLink(false), 2000);
-      },
-      (message) => showError('Copy failed', message),
-    );
-  };
-
-  return (
-    <div className="space-y-5">
-      <div className="space-y-2">
-        <p className="text-sm text-input-placeholder">
-          Share the direct link or install the widget script on your site.
-        </p>
-        <div className="glass-panel rounded-xl px-4 py-3">
-          <div className="flex flex-col gap-3 sm:flex-row sm:items-center sm:justify-between">
-            <div className="min-w-0">
-              <p className="text-xs font-medium uppercase tracking-widest text-input-placeholder">Public link</p>
-              <a
-                href={publicUrl}
-                target="_blank"
-                rel="noreferrer"
-                className="mt-1 block truncate text-sm font-medium text-input-text underline decoration-line-glass/50 underline-offset-4 hover:decoration-input-text"
-              >
-                {publicUrl}
-              </a>
-            </div>
-            <Button type="button" variant="secondary" size="sm" onClick={handleCopyLink}>
-              {copiedLink ? 'Copied' : 'Copy link'}
-            </Button>
-          </div>
-        </div>
-      </div>
-
-      <div className="space-y-3">
-        <p className="text-sm text-input-placeholder">
-          Paste this script into your site&apos;s <code>&lt;head&gt;</code> or before <code>&lt;/body&gt;</code> to load this intake flow.
-        </p>
-        <div className="relative group">
-          <pre className="bg-elevation-2 overflow-x-auto rounded-xl border border-line-glass/30 p-4 pr-20 text-sm font-mono text-input-text">
-            {snippet}
-          </pre>
-          <Button
-            type="button"
-            variant="secondary"
-            size="sm"
-            onClick={handleCopy}
-            className="absolute right-3 top-3 opacity-0 transition-opacity group-hover:opacity-100 group-focus-within:opacity-100 focus-visible:opacity-100"
-          >
-            {copied ? 'Copied' : 'Copy'}
-          </Button>
-        </div>
-      </div>
-    </div>
-  );
-}
+// Embed UI lives in src/features/intake/components/EmbedCodeBlock.tsx
 
 type StatPillProps = {
   label: string;
@@ -975,6 +834,14 @@ function TemplateCard({
   onArchive,
 }: TemplateCardProps) {
   const { showSuccess, showError } = useToastContext();
+  const [openEmbedDialog, setOpenEmbedDialog] = useState(false);
+  const embedTextareaRef = useRef<HTMLTextAreaElement | null>(null);
+  useEffect(() => {
+    if (openEmbedDialog && embedTextareaRef.current) {
+      embedTextareaRef.current.focus();
+      embedTextareaRef.current.select();
+    }
+  }, [openEmbedDialog]);
   const questionPreview = template.fields
     .map((field) => getFieldCanvasQuestion(field))
     .filter((question) => question.trim().length > 0)
@@ -1031,11 +898,7 @@ function TemplateCard({
               </DropdownMenuItem>
               <DropdownMenuItem
                 onSelect={() => {
-                  copyTextToClipboard(
-                    embedSnippet,
-                    () => showSuccess('Embed copied', 'The widget snippet is ready to paste.'),
-                    (message) => showError('Copy failed', message),
-                  );
+                  setOpenEmbedDialog(true);
                 }}
               >
                 Copy embed code
@@ -1053,6 +916,45 @@ function TemplateCard({
             </DropdownMenuContent>
           </DropdownMenu>
         </div>
+        {openEmbedDialog ? (
+          <Dialog
+            isOpen={openEmbedDialog}
+            onClose={() => setOpenEmbedDialog(false)}
+            title="Embed code"
+            description="Copy or inspect the embed snippet for this intake template."
+          >
+            <DialogBody>
+              <p className="mb-3 text-sm text-input-placeholder">Paste this script into your site's <code>&lt;head&gt;</code> or before <code>&lt;/body&gt;</code>.</p>
+              <textarea
+                ref={embedTextareaRef}
+                readOnly
+                value={embedSnippet}
+                className="w-full resize-none rounded-md border border-line-glass/20 bg-surface-ground p-3 font-mono text-sm text-input-text"
+                rows={6}
+                aria-label="Embed snippet"
+              />
+            </DialogBody>
+            <DialogFooter>
+              <Button
+                type="button"
+                variant="secondary"
+                onClick={() => {
+                  copyTextToClipboard(
+                    embedSnippet,
+                    () => {
+                      showSuccess('Embed copied', 'The widget snippet is ready to paste.');
+                      setOpenEmbedDialog(false);
+                    },
+                    (message) => showError('Copy failed', message),
+                  );
+                }}
+              >
+                Copy embed
+              </Button>
+              <Button type="button" variant="ghost" onClick={() => setOpenEmbedDialog(false)}>Close</Button>
+            </DialogFooter>
+          </Dialog>
+        ) : null}
 
         <div className="mt-5 space-y-2">
           {questionPreview.map((question, index) => (
